[RFC] Add formatError method option
Specifying a custom message for every field because the default message is a bit too little verbose for me is pretty tedious. This patch introduces a new option to `createRateLimitDirective` called `formatError` (same as in graphql-js) that gets a bunch of parameters as arguments and can return a custom error message.